+++ Examples/Digit/include/DigitDynamicsConstraints.h
@@ -58,12 +58,12 @@ const std::string independentJointNames[NUM_INDEPENDENT_JOINTS] =
 
 class DigitDynamicsConstraints : public DynamicsConstraints {
 public:
-    using Model = pinocchio::ModelTpl<float>;
-    using Data = pinocchio::DataTpl<float>;
-    using Vec3 = Eigen::Vector3f;
-    using Mat3 = Eigen::Matrix3f;
-    using VecX = Eigen::VectorXf;
-    using MatX = Eigen::MatrixXf;
+    using Model = pinocchio::ModelTpl<double>;
+    using Data = pinocchio::DataTpl<double>;
+    using Vec3 = Eigen::Vector3d;
+    using Mat3 = Eigen::Matrix3d;
+    using VecX = Eigen::VectorXd;
+    using MatX = Eigen::MatrixXd;
 
     // Constructor
     DigitDynamicsConstraints() = default;
