+++ Examples/Digit/src/DigitSingleStepPeriodicityConstraints.cpp
@@ -99,7 +99,7 @@ void DigitSingleStepPeriodicityConstraints::compute(const VecX& z,
 
     if (compute_derivatives) {
         // compute derivatives with gravity turned off, we just need prnea_pq here
-        const float original_gravity = dcidPtr_->modelPtr_->gravity.linear()(2);
+        const double original_gravity = dcidPtr_->modelPtr_->gravity.linear()(2);
         dcidPtr_->modelPtr_->gravity.linear()(2) = 0;
         pinocchio::computeRNEADerivatives(*(dcidPtr_->modelPtr_), *(dcidPtr_->dataPtr_), 
                                           q_minus, zeroVec, v_plus - v_minus,
@@ -152,14 +152,14 @@ void DigitSingleStepPeriodicityConstraints::compute(const VecX& z,
     // (5) contact constraints
     VecX contactLambda = lambda.tail(6);
 
-    float contact_force         = contactLambda(2);
-    float friction_force_sq     = pow(contactLambda(0), 2) + pow(contactLambda(1), 2);
-    float max_friction_force_sq = pow(fp.mu * contact_force, 2);
-    float max_moment_z_sq       = pow(fp.gamma * contact_force, 2);
-    float mx_lower_limit        = -fp.Lx * contact_force;
-    float mx_upper_limit        = fp.Lx * contact_force;
-    float my_lower_limit        = -fp.Ly * contact_force;
-    float my_upper_limit        = fp.Ly * contact_force;
+    double contact_force         = contactLambda(2);
+    double friction_force_sq     = pow(contactLambda(0), 2) + pow(contactLambda(1), 2);
+    double max_friction_force_sq = pow(fp.mu * contact_force, 2);
+    double max_moment_z_sq       = pow(fp.gamma * contact_force, 2);
+    double mx_lower_limit        = -fp.Lx * contact_force;
+    double mx_upper_limit        = fp.Lx * contact_force;
+    double my_lower_limit        = -fp.Ly * contact_force;
+    double my_upper_limit        = fp.Ly * contact_force;
 
     // (1) positive contact force
     g5(0) = contact_force;
